backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
pydantic>=1.8.0,<1.9.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
implicit>=0.6.0
python-multipart>=0.0.5,<0.0.6
aioredis>=2.0.0,<2.1.0
python-dotenv>=0.19.0,<0.20.0
pandas>=1.4.0,<1.5.0
psycopg2-binary>=2.9.0,<2.10.0
pytest>=7.0.0,<7.1.0
pytest-asyncio>=0.18.0,<0.19.0
httpx>=0.23.0,<0.24.0